mc-api/requirements.txt
Flask==2.1.3
Werkzeug==2.2.3
psycopg2-binary==2.9.6
gunicorn==20.1.0
Flask-Bcrypt==1.0.1
PyJWT==2.6.0
Flask-Mail==0.9.1
Flask-SocketIO==5.3.3
redis==4.3.5
eventlet==0.30.2
flask-cors==3.0.10
cryptography==3.4.7
loguru
watchdog
faker
humanize==4.0.0
